+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.7.0]
+### Changed
+* Optimized cmake
+* **Secondary alignments are filtered now**
+
+### Added
+* Output UMIs with only exonic or only intronic reads 
+
 ## [0.6.8] - 2017-09-22
 ### Added
 * Filtration of reads by barcode quality ("*TagsSearch/Processing/min_barcode_quality*" and